+++ src/components/Table/index.ts
@@ -2,6 +2,8 @@ export { Table } from './Table.component';
 export { TableFiltersDropdown } from './TableFiltersDropdown.component';
 export type {
   TableColumn,
+  TableValueColumn,
+  TableCustomColumn,
   TableProps,
   TableFilters,
   FilterValue,