+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export { renderIcon } from './renderIcon';
 export type { IconType } from './renderIcon';
 export type { ComponentSizeProps } from './sizes';
+export type { RowKey } from './rowKey';